+++ emmyLuaDocs/bait.lua
@@ -12,6 +12,11 @@ function bait.catch(name, cb) end
 --- @param cb function
 function bait.catchOnce(name, cb) end
 
+--- Removes the `catcher` for the event with `name`
+--- For this to work, `catcher` has to be the same function used to catch
+--- an event, like one saved to a variable.
+function bait.release() end
+
 --- Throws a hook with `name` with the provided `args`
 --- @param name string
 --- @vararg any