+++ exec.go
@@ -85,7 +85,7 @@ func isExecError(err error) (execError, bool) {
 func runInput(input string, priv bool) {
 	running = true
 	cmdString := aliases.Resolve(input)
-	hooks.Em.Emit("command.preexec", input, cmdString)
+	hooks.Emit("command.preexec", input, cmdString)
 
 	rerun:
 	var exitCode uint8
@@ -101,7 +101,7 @@ func runInput(input string, priv bool) {
 					cmdFinish(0, input, priv)
 					return
 				}
-				input, exitCode, cont, err = handleSh(input)
+				input, exitCode, cont, err = handleSh(cmdString)
 			case "hybridRev":
 				_, _, _, err = handleSh(input)
 				if err == nil {
@@ -112,7 +112,7 @@ func runInput(input string, priv bool) {
 			case "lua":
 				input, exitCode, err = handleLua(cmdString)
 			case "sh":
-				input, exitCode, cont, err = handleSh(input)
+				input, exitCode, cont, err = handleSh(cmdString)
 		}
 	} else {
 		// can only be a string or function so
@@ -140,7 +140,7 @@ func runInput(input string, priv bool) {
 
 	if err != nil {
 		if exErr, ok := isExecError(err); ok {
-			hooks.Em.Emit("command." + exErr.typ, exErr.cmd)
+			hooks.Emit("command." + exErr.typ, exErr.cmd)
 			err = exErr.sprint()
 		}
 		fmt.Fprintln(os.Stderr, err)
@@ -152,6 +152,7 @@ func reprompt(input string) (string, error) {
 	for {
 		in, err := continuePrompt(strings.TrimSuffix(input, "\\"))
 		if err != nil {
+			lr.SetPrompt(fmtPrompt(prompt))
 			return input, err
 		}
 
@@ -220,7 +221,17 @@ func handleLua(cmdString string) (string, uint8, error) {
 	return cmdString, 125, err
 }
 
-func handleSh(cmdString string) (string, uint8, bool, error) {
+func handleSh(cmdString string) (input string, exitCode uint8, cont bool, runErr error) {
+	shRunner := hshMod.Get(rt.StringValue("runner")).AsTable().Get(rt.StringValue("sh"))
+	var err error
+	input, exitCode, cont, runErr, err = runLuaRunner(shRunner, cmdString)
+	if err != nil {
+		runErr = err
+	}
+	return
+}
+
+func execSh(cmdString string) (string, uint8, bool, error) {
 	_, _, err := execCommand(cmdString, true)
 	if err != nil {
 		// If input is incomplete, start multiline prompting
@@ -544,5 +555,5 @@ func cmdFinish(code uint8, cmdstr string, private bool) {
 	// using AsValue (to convert to lua type) on an interface which is an int
 	// results in it being unknown in lua .... ????
 	// so we allow the hook handler to take lua runtime Values
-	hooks.Em.Emit("command.exit", rt.IntValue(int64(code)), cmdstr, private)
+	hooks.Emit("command.exit", rt.IntValue(int64(code)), cmdstr, private)
 }